- Updated project with tests and many other little things

+++ src/CloudSalesSystem.Application/CloudComputingService/Models/SoftwareService.cs
@@ -1,4 +1,4 @@
 ﻿namespace CloudSalesSystem.Application.CloudComputingService.Models;
 
 public record SoftwareServiceResponse(IReadOnlyList<SoftwareService> Items, int Page, int PageSize, int Total);
-public record SoftwareService(Guid Id, string Name, string Description, int AvailableLicenses, int Price, DateTime ValidToDate);
+public record SoftwareService(Guid Id, string Name, string Description, int AvailableLicenses, decimal Price, DateTime ValidToDate);
